work on batch actions

+++ OmasApi/Controllers/OrderBatchesController.cs
@@ -3,8 +3,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Internal;
 using OmasApi.Controllers.Middleware;
 using OmasApi.Data;
+using OmasApi.Models;
 
 namespace OmasApi.Controllers
 {
@@ -30,16 +32,26 @@ namespace OmasApi.Controllers
 
 
         [HttpGet("{id}")]
-        public async Task<OrderBatch> Get(int id)
+        public async Task<OrderBatchModel> Get(int id)
         {
-            var batch = await _db.OrderBatches.FindAsync(id);
+            var batch = await _db.OrderBatches.Include(b => b.OrderItems).SingleOrDefaultAsync(b => b.BatchId == id);
 
-            if (batch== null)
+            if (batch == null)
             {
                 throw new NotFoundException($"Batch '{id}' does not exist");
             }
 
-            return batch;
+            var model = new OrderBatchModel
+            {
+                BatchId = batch.BatchId,
+                OrderDate = batch.OrderDate,
+                DeliveryDate = batch.DeliveryDate,
+                IsOpen = batch.IsOpen,
+                CustomerCount = batch.OrderItems.Select(i => i.UserId).Distinct().Count(),
+                Total = batch.OrderItems.Sum(i => i.Price * i.Quantity)
+            };
+
+            return model;
         }
 
         [HttpPost]
@@ -78,7 +90,12 @@ namespace OmasApi.Controllers
                 throw new ReferentialIntegrityException($"Batch '{id}' contains orders and cannot be deleted");
             }
 
-            var batch = await Get(id);
+            var batch = await _db.OrderBatches.FindAsync(id);
+
+            if (batch == null)
+            {
+                throw new NotFoundException($"Batch '{id}' does not exist");
+            }
 
             _db.OrderBatches.Remove(batch);
             await _db.SaveChangesAsync();
